Something kind of resembling a score.
Crppy lerp.

Some entry for the tilemap. Kinda of looks crap because the tilemap isn't clipped right at the moment.

Add screen shake when tile is full.

Pulsing colour and filling the progress tile works.
NOTE: Any colours pulled from a palette that was auto-generated will be in next format, e.g. lowest bit of blue in top byte.
This will convert from one to the other and back again during blend operations.
Moved a bunch of variables from global -> globals, but not inside the globals structure. It means for a less cluttered code.
Added utility function to set a colour.