Make sure to use top bit for Y for player.
Fixed some play area tiling issues. Aligned tilemap and player to center of screen.

Move files around a bit. The main loop was getting a bit large and it made sense to split out between Initialize and Reset on various subsystems. The resets don't do much more than the initialize at the moment. But this is likely to change.

Hide tilemap until it's fully initialized.

Fixed the particle problem. It's data was not getting aligned to a 16 byte boundary. This caused issues with the 'inc l' within the update loop. This is to avoid doing inc hl