Fixup reset sequence so we don't zero out the score.
Invalidate the particle page when restoring from other than the render loop. This caused sporadic corruption in other places due to the 'cache' page number being valid, yet paged to the wrong place.

Add coin & tile count display.
Font palette was getting corrupted due to the tile palette only being 16 entries. This needed to be 256, so that there was sufficient space to copy that palette for the tile blocks.